+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `--batch-size` option to `varlogcli` to set the size of append batch. (#VARLOG-653/570)
 - Added metrics about the delay between reports and commits in the metadata repository. (#VARLOG-654/#571)
 - Added metrics about the various performance indices in the storage node. (#VARLOG-656/#573)
+- Added `--ballast-size` option to storage node. (#VARLOG-669/#585)
 
 ### Changed
 - Fixed wrong URL in `CHANGELOG.md`. (#VARLOG-652/#569)